Collected all the constants into one file
